requirements.txt
certifi==2019.3.9
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2==2.7.6.1
python-dotenv==0.10.2
six==1.12.0
Werkzeug==0.15.4
